LayoutTests/webgpu/whlsl/use-undefined-variable-2-expected.txt

CONSOLE MESSAGE: GPUDevice.createComputePipeline(): WHLSL compile error: 4:25: Cannot find the variable declaration.
PASS
PASS successfullyParsed is true

TEST COMPLETE
